refactor: extract logic from month calendar page into dedicated component
