+++ src/components/calendar/index.ts
@@ -1,2 +1,1 @@
-export { default as MonthCalendarGrid } from './MonthCalendarGrid';
-export { default as MonthCalendarHeader } from './MonthCalendarHeader';
+export { default as MonthCalendar } from './MonthCalendar';
