+++ src/pages/MonthCalendarPage.tsx
@@ -1,27 +1,9 @@
-import { CalendarDate, GregorianCalendar } from '@internationalized/date';
-import {
-  endOfDay,
-  endOfWeek,
-  endOfMonth,
-  startOfDay,
-  startOfWeek,
-  startOfMonth,
-  startOfToday,
-} from 'date-fns';
+import { GregorianCalendar } from '@internationalized/date';
 import React from 'react';
-import { useCalendar, useDateFormatter } from 'react-aria';
-import { useParams } from 'react-router';
+import { useCalendar } from 'react-aria';
 import { useCalendarState } from 'react-stately';
 
-import { MonthCalendarGrid, MonthCalendarHeader } from '@components';
-import { useUser } from '@hooks';
-import type { OccurrenceFilters } from '@models';
-import {
-  useHabits,
-  useTraits,
-  useOccurrences,
-  useOccurrenceActions,
-} from '@stores';
+import { MonthCalendar } from '@components';
 
 const createCalendar = (identifier: string) => {
   switch (identifier) {
@@ -33,114 +15,18 @@ const createCalendar = (identifier: string) => {
   }
 };
 
-type MonthCalendarProps = {
+type MonthCalendarPageProps = {
   locale: string;
 };
 
-const MonthCalendar = ({ locale }: MonthCalendarProps) => {
-  const { user } = useUser();
-  const occurrences = useOccurrences();
-  const habits = useHabits();
-  const traits = useTraits();
-  const { clearOccurrences, fetchOccurrences } = useOccurrenceActions();
-  const [filters, setFilters] = React.useState<OccurrenceFilters>({
-    habitIds: new Set(),
-    traitIds: new Set(),
-  });
-  const [fetchedMonthYear, setFetchedMonthYear] = React.useState<string>('');
+const MonthCalendarPage = ({ locale }: MonthCalendarPageProps) => {
   const calendarState = useCalendarState({
     createCalendar,
     firstDayOfWeek: 'mon',
     isReadOnly: true,
     locale,
   });
-  const formatter = useDateFormatter({
-    month: 'long',
-    timeZone: calendarState.timeZone,
-  });
   const { calendarProps, title } = useCalendar({}, calendarState);
-  const params = useParams();
-  const months = [];
-
-  const numMonths = calendarState.focusedDate.calendar.getMonthsInYear(
-    calendarState.focusedDate
-  );
-
-  for (let i = 1; i <= numMonths; i++) {
-    const date = calendarState.focusedDate.set({ month: i });
-    months.push(formatter.format(date.toDate(calendarState.timeZone)));
-  }
-
-  const derivedFilters = React.useMemo(() => {
-    return {
-      habitIds: new Set(
-        Object.values(habits).map((habit) => {
-          return habit.id.toString();
-        })
-      ),
-      traitIds: new Set(
-        Object.values(traits).map((trait) => {
-          return trait.id.toString();
-        })
-      ),
-    };
-  }, [habits, traits]);
-
-  React.useEffect(() => {
-    if (!filters.habitIds.size || !filters.traitIds.size) {
-      setFilters(derivedFilters);
-    }
-  }, [derivedFilters, filters.habitIds.size, filters.traitIds.size]);
-
-  React.useEffect(() => {
-    if (!user) {
-      if (occurrences.length) {
-        clearOccurrences();
-      }
-
-      return;
-    }
-
-    const currentMonth = startOfMonth(startOfToday());
-
-    const {
-      day = currentMonth.getDate(),
-      month = currentMonth.getMonth() + 1,
-      year = currentMonth.getFullYear(),
-    } = params;
-
-    const paramsDate = new Date(Number(year), Number(month) - 1, Number(day));
-
-    const nextFocusedDate = new CalendarDate(
-      paramsDate.getFullYear(),
-      paramsDate.getMonth() + 1,
-      paramsDate.getDate()
-    );
-
-    const hasFocusedDateChanged =
-      calendarState.focusedDate.toString() !== nextFocusedDate.toString();
-
-    const currentMonthYear = `${Number(year)}-${Number(month)}`;
-    const hasNotFetchedThisMonth = fetchedMonthYear !== currentMonthYear;
-
-    if (hasFocusedDateChanged || hasNotFetchedThisMonth) {
-      const rangeStart = startOfDay(startOfWeek(startOfMonth(paramsDate)));
-      const rangeEnd = endOfDay(endOfWeek(endOfMonth(paramsDate)));
-
-      void fetchOccurrences([+rangeStart, +rangeEnd]);
-      setFetchedMonthYear(currentMonthYear);
-
-      calendarState.setFocusedDate(nextFocusedDate);
-    }
-  }, [
-    fetchedMonthYear,
-    params,
-    calendarState,
-    user,
-    fetchOccurrences,
-    clearOccurrences,
-    occurrences.length,
-  ]);
 
   const [activeMonthLabel, activeYear] = title.split(' ');
 
@@ -152,26 +38,13 @@ const MonthCalendar = ({ locale }: MonthCalendarProps) => {
       <title>
         {`${activeMonthLabel.slice(0, 3)} ${activeYear} | Habitrack Calendar`}
       </title>
-      <MonthCalendarHeader
-        months={months}
-        filters={filters}
-        activeYear={activeYear}
-        onFilterChange={setFilters}
-        activeMonthLabel={activeMonthLabel}
-      />
-      <MonthCalendarGrid
+      <MonthCalendar
         state={calendarState}
-        activeYear={Number(activeYear)}
-        activeMonthIndex={months.indexOf(activeMonthLabel)}
-        occurrences={occurrences.filter((occurrence) => {
-          return (
-            filters.habitIds.has(occurrence.habitId.toString()) &&
-            filters.traitIds.has(occurrence.habit?.trait?.id.toString() || '')
-          );
-        })}
+        activeYear={activeYear}
+        activeMonthLabel={activeMonthLabel}
       />
     </div>
   );
 };
 
-export default MonthCalendar;
+export default MonthCalendarPage;
